test(runtime): share the in-memory artifact store across compact tests
Pure refactor: the loader unit tests and the backend replay e2e each
carried their own copy of the same Map-backed store; both now import a
single helper (no .test. suffix, so the runner glob skips it).
